pyproject.toml
[project]
name = "biometric-mlops"
version = "0.1.0"
description = "Production-grade MLOps infrastructure for multimodal biometric recognition (iris + fingerprint)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pydantic>=2.5.0",
    "numpy>=1.26.0",
    "pyarrow>=14.0.0",
    "structlog>=24.1.0",
    "pillow>=10.0.0",
    "hydra-core>=1.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "mlflow>=2.10.0",
    "matplotlib>=3.8.0",
    "pip-audit>=2.6.0",
    "bandit[toml]>=1.7.0",
    "pre-commit>=4.0.0",
]
ray = ["ray[data]>=2.9.0"]
mlflow = ["mlflow>=2.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biometric"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "pyarrow",
    "pyarrow.*",
    "torchvision",
    "torchvision.transforms",
    "torchvision.transforms.*",
    "ray",
    "ray.data",
    "ray.data.*",
]
ignore_missing_imports = true


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
pythonpath = ["src"]
markers = [
    "performance: marks tests as performance benchmarks (deselect with '-m \"not performance\"')",
]

[tool.coverage.run]
source = ["src/biometric"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
